JPA Server side test fixes - part 1 This is first part of the fix for JPA test apps executed on JEE server environment. Main changes there: - eclipselink-testbuild-plugin upgrade to 1.1.2 with <libs> support - (JpaEntityManager)entityManager -> entityManager.unwrap(JpaEntityManager) as server should provide proxy class - em.getTransaction().begin() -> beginTransaction(em) to allow JPA test FW identify if test is executed in JSE or JEE environment and call proper code

+++ jpa/eclipselink.jpa.testapps/jpa.test.jpaadvancedproperties/src/test/java/org/eclipse/persistence/testing/tests/jpa/jpaadvancedproperties/JPAAdvPropertiesTest.java
@@ -302,27 +302,30 @@ public class JPAAdvPropertiesTest extends JUnitTestCase {
     }
 
     public void testLoginEncryptorProperty() {
-        EntityManager em = createEntityManager();
-        try {
-            //Create new customer
-            beginTransaction(em);
-            Customer customer = ModelExamples.customerExample1();
-            em.persist(customer);
-            em.flush();
-            Integer customerId = customer.getCustomerId();
-            commitTransaction(em);
+        //This test is valid for JSE environment
+        if (! isOnServer()) {
+            EntityManager em = createEntityManager();
+            try {
+                //Create new customer
+                beginTransaction(em);
+                Customer customer = ModelExamples.customerExample1();
+                em.persist(customer);
+                em.flush();
+                Integer customerId = customer.getCustomerId();
+                commitTransaction(em);
 
-            customer = em.find(org.eclipse.persistence.testing.models.jpa.jpaadvancedproperties.Customer.class, customerId);
-            //Purge it
-            beginTransaction(em);
-            em.remove(customer);
-            commitTransaction(em);
+                //Purge it
+                beginTransaction(em);
+                customer = em.find(org.eclipse.persistence.testing.models.jpa.jpaadvancedproperties.Customer.class, customerId);
+                em.remove(customer);
+                commitTransaction(em);
 
-            assertNotNull(customer);
-            assertTrue("CustomizedEncryptor.encryptPassword() method wasn't called.", CustomizedEncryptor.encryptPasswordCounter > 0);
-            assertTrue("CustomizedEncryptor.decryptPassword() method wasn't called.", CustomizedEncryptor.decryptPasswordCounter > 0);
-        } finally {
-            closeEntityManager(em);
+                assertNotNull(customer);
+                assertTrue("CustomizedEncryptor.encryptPassword() method wasn't called.", CustomizedEncryptor.encryptPasswordCounter > 0);
+                assertTrue("CustomizedEncryptor.decryptPassword() method wasn't called.", CustomizedEncryptor.decryptPasswordCounter > 0);
+            } finally {
+                closeEntityManager(em);
+            }
         }
     }
 }
